Include socat workaround
As a workaround to fool the dispatcher to connect to the emulator in an external container, there is a workaround using socat that tunnels the localhost inside the container to another address and port where the emulator is listening. This workaround will disappear once the dispatcher is refactored not to expect the emulator to be local.